more fixes to buffers and chars

+++ UV98_868_220430_D4/MAIN/CRC16.C
@@ -64,7 +64,7 @@ void GetCrc16(const uchar *pData)
 */
     
 // Calculates the 16-bit CRC of data of a given length.
-void GetCrc16_LEN(const uchar *pData,uchar nLength)
+void GetCrc16_LEN(const uchar *pData,uint nLength)
 {
     unsigned short fcs = 0xffff;    // FCS Initialization
    	unsigned short result;
